Update SAN builder to handle multiple attackers
If two of the same piece (e.g. two knights) can move the same square, the SAN builder adds the file to the SAN. So Knight on f3 moving to d2 would go from Nd2 to Nfd2. If they are on the same file then the rank is used so N3d2.